+++ crates/astria-sequencer/src/app/tests_app/mod.rs
@@ -641,7 +641,6 @@ async fn app_prepare_proposal_cometbft_max_bytes_overflow_ok() {
     let height = fixture.block_height().await.increment();
 
     // create txs which will cause cometBFT overflow
-
     let tx_pass = fixture
         .checked_tx_builder()
         .with_action(RollupDataSubmission {