+++ crates/astria-sequencer-relayer/src/relayer/builder.rs
@@ -29,6 +29,7 @@ pub(crate) struct Builder {
     pub(crate) relayer_shutdown_token: tokio_util::sync::CancellationToken,
     pub(crate) sequencer_chain_id: String,
     pub(crate) celestia_chain_id: String,
+    pub(crate) celestia_default_min_gas_price: f64,
     pub(crate) celestia_app_grpc_endpoint: String,
     pub(crate) celestia_app_key_file: String,
     pub(crate) cometbft_endpoint: String,
@@ -46,6 +47,7 @@ impl Builder {
             relayer_shutdown_token,
             sequencer_chain_id,
             celestia_chain_id,
+            celestia_default_min_gas_price,
             celestia_app_grpc_endpoint,
             celestia_app_key_file,
             cometbft_endpoint,
@@ -77,8 +79,14 @@ impl Builder {
                 .wrap_err("failed parsing provided celestia app grpc endpoint as Uri")?;
             let celestia_keys = CelestiaKeys::from_path(celestia_app_key_file)
                 .wrap_err("failed to get celestia keys from file")?;
-            CelestiaClientBuilder::new(celestia_chain_id, uri, celestia_keys, state.clone())
-                .wrap_err("failed to create celestia client builder")?
+            CelestiaClientBuilder::new(
+                celestia_chain_id,
+                celestia_default_min_gas_price,
+                uri,
+                celestia_keys,
+                state.clone(),
+            )
+            .wrap_err("failed to create celestia client builder")?
         };
 
         Ok(super::Relayer {